(fix) Handle error when failing to checkout out a SHA
We believe that a good (simple) convention is for a module's version'
always to be a tag or a branch name rather than a particular SHA.

If the attempt to `git checkout` a particular SHA results in a
'fatal: reference is not a tree:' error, we catch that error and print
out a warning.

In the near future we will start re-raising the error to enforce the
use of tag and branch names only.